requirements.txt
streamlit
opencv-python-headless
matplotlib       
numpy
imutils
easyocr
Pillow
pickle-mixin